donation complete

+++ dao/project_dao.go
@@ -22,7 +22,7 @@ func GetProjectByName(name string) (project *model.Project,err error){
 }
 
 //根据userid获取所有项目
-func ListProjectsByUserId(userId int64)(projects *[]model.Project,err error){
+func ListProjectsByUserId(userId uint32)(projects *[]model.Project,err error){
 	projects = &[]model.Project{}
 	err = util.DB.Debug().Where("id IN(?)",
 		util.DB.Debug().Table("member").Select("project_id").Where("user_id=?",userId).SubQuery()).Find(projects).Error
